즐겨찾기 기능 구현

+++ CafeStar/cafe-star/src/redux/states/favoriteCafeList.js
@@ -10,24 +10,22 @@ export const favoriteCafeListSlice = createSlice({
       state.value = [...action.payload];
     },
     pushFavoriteCafe: (state, action) => {
-      fetch('http://localhost:5000/favorite-cafes/push', {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(action.payload),
-      })
-        .then((response) => response.json())
-        .then(data => state.value.push(data));
+      if (state.value.findIndex(item => item.id === action.payload.id) === -1) {
+        fetch('http://localhost:5000/favorite-cafes/push', {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify(action.payload),
+        });
+      }
     },
     removeFavoriteCafe: (state, action) => {
-      state.value = state.value.splice(state.value.findIndex(item => item.id === action.payload), 1);
-
-      fetch('http://localhost:5000/favorite-cafes/delete?id=' + action.payload, {
-        method: "DELETE",
-      })
-      .then((response) => response.json())
-      .then(data => console.log(data));
+      if (state.value.findIndex(item => item.id === action.payload.id) !== -1) {
+        fetch('http://localhost:5000/favorite-cafes/delete?id=' + action.payload.id, {
+          method: "DELETE",
+        });
+      }
     },
   },
 });
